Enhance(player): Emby-style quality picker + playback info + subtitle fixes
Backend:
- Add video_fps, video_profile, video_level to media_files (migration 018)
- Add sample_rate to audio_tracks (migration 018)
- Expose new fields in PlaybackInfoResponse + AudioTrackInfo
- Add Subdl subtitle provider with episode→series metadata resolution
- Add admin playback settings (direct_play mode, auto-subtitles)
- Add playback mode query param (?pm=) for DirectStream vs DirectPlay
- ffprobe: parse r_frame_rate, profile, level, sample_rate

Frontend:
- Emby-style quality picker with submenu navigation (24 presets)
- Playback Info overlay redesign (not toggle, overlay with close button)
- Display video profile/level/fps + audio bitrate/sample_rate
- Client-side fps fallback for files not yet rescanned
- Subtitle search: Subdl provider + ISO 639-2→1 mapping
- Settings page: Subdl, playback mode, auto-subtitle config
- Fix video restart on subtitle/resolution change (read position from store)
- Fix overlay menus not hiding on play toggle

+++ backend/internal/model/user.go
@@ -40,9 +40,9 @@ type UserData struct {
 	UpdatedAt    string   `json:"updated_at"`
 
 	// JOIN fields (populated by queries with media)
-	MediaTitle    string `json:"media_title,omitempty"`
-	MediaPoster   string `json:"media_poster,omitempty"`
-	MediaDuration int    `json:"media_duration,omitempty"`
+	MediaTitle    string  `json:"media_title,omitempty"`
+	MediaPoster   string  `json:"media_poster,omitempty"`
+	MediaDuration float64 `json:"media_duration,omitempty"`
 }
 
 // UserSeriesData represents series-level favorite/rating
